requirements.txt
python-telegram-bot[ext]==21.0.1
yt-dlp==2024.04.09
python-dotenv==1.0.1
Flask==3.0.3
gunicorn==22.0.0
uvicorn==0.29.0
wsgi_adapter==0.3.0
